+++ tests/rules/RulesEngine_Auror.test.ts
@@ -37,6 +37,7 @@ describe("RulesEngine", function() {
         enemy = TestData.buildDefaultEnemy();
         facts = {
             wizard: wizard,
+            lowestHPWizard: wizard, 
             highestPriorityAvailableEnemy: enemy,
             allWizards: [wizard],
             allActiveEnemies: [enemy],